+++ packages/terra-application-links/CHANGELOG.md
@@ -3,9 +3,20 @@ ChangeLog
 
 Unreleased
 -----------------
+
+4.9.0 - (July 26, 2018)
+------------------
+### Changed
+* Minor dependency updates
+
+4.8.0 - (July 25, 2018)
+------------------
 ### Added
 * Added `sv` and `sv-SE` translations
 * Expanded support for navigational paths.
+
+### Changed
+* Add .module extension to SCSS file(s)
 
 4.7.0 - (July 10, 2018)
 ------------------