+++ packages/terra-brand-footer/CHANGELOG.md
@@ -3,9 +3,15 @@ ChangeLog
 
 Unreleased
 -----------------
+
+1.7.0 - (July 25, 2018)
+------------------
 ### Added
 * Section headers can be put in the nav bar
 * The navigation bar can be displayed vertically
+
+### Changed
+* Add .module extension to SCSS file(s)
 
 1.6.0 - (July 10, 2018)
 ------------------