+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -3,8 +3,19 @@ ChangeLog
 
 Unreleased
 -----------------
+
+1.3.0 - (July 26, 2018)
+------------------
+### Changed
+* Minor dependency updates
+
+1.2.0 - (July 25, 2018)
+------------------
 ### Fixed
 * Fix webpack issue due to referencing src in the docs examples and tests instead of a relative path which would end up hitting lib which has been transpiled.
+
+### Changed
+* Add .module extension to SCSS file(s)
 
 1.1.0 - (July 3, 2018)
 ------------------