+++ packages/terra-infinite-list/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.8.0 - (July 25, 2018)
+------------------
+### Changed
+* Add .module extension to SCSS file(s)
+
 1.7.0 - (July 10, 2018)
 ------------------
 ### Changed