+++ packages/terra-layout/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+2.11.0 - (July 26, 2018)
+------------------
+### Changed
+* Updated to use terra-responsive-element v3
+
+2.10.0 - (July 25, 2018)
+------------------
+### Changed
+* Add .module extension to SCSS file(s)
+
 2.9.0 - (July 3, 2018)
 ------------------
 ### Added