+++ packages/terra-popup/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+4.9.0 - (July 26, 2018)
+------------------
+### Changed
+* Minor dependency updates
+
+4.8.0 - (July 25, 2018)
+------------------
+### Changed
+* Add .module extension to SCSS file(s)
+
 4.7.0 - (July 10, 2018)
 ------------------
 ### Changed